Moved: getHorizontal, getVertical, calcOffset, default labels, label ID setter, lots of time stuff
