+++ src/xDate.js
@@ -4,7 +4,17 @@
  */
 
 import * as d3 from 'd3';
-import { getAxis, convertToPointScale } from './utils';
+import {
+    convertToPointScale,
+    getAxis,
+    getDefaultXAxisLabel,
+    getTimeTickFormat,
+    getTimeTicks,
+    getTimeTicksMinor,
+    getVertical,
+    getHorizontal,
+    setLabelIds,
+} from './utils';
 
 export default function xaxisDate() {
     let banding;
@@ -39,7 +49,7 @@ export default function xaxisDate() {
             scale = convertToPointScale(scale);
             xAxis
                 .tickSize(tickSize)
-                .tickFormat(tickFormat(interval))
+                .tickFormat(getTimeTickFormat(interval, { fullYear, scale }))
                 .scale(scale);
             xAxis.tickValues(
                 scale.domain().filter((d, i) => {
@@ -56,10 +66,10 @@ export default function xaxisDate() {
         } else {
             xAxis
                 .tickSize(tickSize)
-                // .ticks(getTicks(interval))
-                .tickFormat(tickFormat(interval))
+                // .ticks(getTimeTicks(interval))
+                .tickFormat(getTimeTickFormat(interval, { fullYear, scale }))
                 .scale(scale);
-            let newTicks = scale.ticks(getTicks(interval));
+            let newTicks = scale.ticks(getTimeTicks(interval));
             const dayCheck = scale.domain()[0].getDate();
             const monthCheck = scale.domain()[0].getMonth();
             if (dayCheck !== 1 && monthCheck !== 0) {
@@ -88,7 +98,7 @@ export default function xaxisDate() {
         } else {
             xMinor
                 .tickSize(minorTickSize)
-                .ticks(getTicksMinor(interval))
+                .ticks(getTimeTicksMinor(interval))
                 .tickFormat('')
                 .scale(scale);
         }
@@ -121,12 +131,7 @@ export default function xaxisDate() {
         }
 
         if (frameName) {
-            xLabel
-                .selectAll('.axis.xAxis text')
-                .attr('id', `${frameName}xLabel`);
-            xLabel
-                .selectAll('.axis.xAxis line')
-                .attr('id', `${frameName}xTick`);
+            setLabelIds({ selection: xLabel, axis: 'x', frameName });
             if (minorAxis) {
                 xLabelMinor
                     .selectAll('.axis.xAxis line')
@@ -135,42 +140,26 @@ export default function xaxisDate() {
         }
 
         if (label) {
-            const defaultLabel = {
-                tag: label.tag,
-                hori: label.hori || 'middle',
-                vert: label.vert || 'bottom',
-                anchor: label.anchor || 'middle',
-                rotate: label.rotate || 0,
-            };
+            const defaultLabel = getDefaultXAxisLabel(label);
 
             const axisLabel = parent.append('g').attr('class', 'axis xAxis');
-            const calcOffset = () => {
-                if (tickSize > 0 && tickSize < rem) {
-                    return tickSize + (rem * 0.8); // prettier-ignore
-                }
-                return rem * 0.9;
-            };
-            const getVertical = (axisAlign, vertAlign) =>
-                ({
-                    toptop: 0 - rem,
-                    topmiddle: 0,
-                    topbottom: 0 + rem,
-                    bottomtop: plotHeight,
-                    bottommiddle: plotHeight + calcOffset(),
-                    bottombottom: plotHeight + calcOffset() + (rem * 1.1), // prettier-ignore
-                }[axisAlign + vertAlign]);
-
-            const getHorizontal = hori =>
-                ({
-                    left: plotWidth - plotWidth,
-                    middle: plotWidth / 2,
-                    right: plotWidth,
-                }[hori]);
 
             axisLabel
                 .append('text')
-                .attr('y', getVertical(align, defaultLabel.vert))
-                .attr('x', getHorizontal(defaultLabel.hori))
+                .attr(
+                    'y',
+                    getVertical({
+                        axisAlign: align,
+                        vertAlign: defaultLabel.vert,
+                        plotHeight,
+                        rem,
+                        tickSize,
+                    }),
+                )
+                .attr(
+                    'x',
+                    getHorizontal({ hori: defaultLabel.hori, plotWidth }),
+                )
                 .text(defaultLabel.tag);
 
             const text = axisLabel.selectAll('text');
@@ -198,7 +187,7 @@ export default function xaxisDate() {
                 }))
                 .filter((d, i) => i % 2 === 0);
 
-            console.log('bands', bands);
+            // console.log('bands', bands);
 
             bandHolder
                 .selectAll('rect')
@@ -214,172 +203,6 @@ export default function xaxisDate() {
         xLabel.selectAll('.domain').remove();
     }
 
-    function getTicks(intvl) {
-        return {
-            century: d3.timeYear.every(100),
-            jubilee: d3.timeYear.every(50),
-            decade: d3.timeYear.every(10),
-            lustrum: d3.timeYear.every(5),
-            years: d3.timeYear.every(1),
-            fiscal: d3.timeYear.every(1),
-            quarters: d3.timeYear.every(1),
-            months: d3.timeMonth.every(1),
-            weeks: d3.timeWeek.every(1),
-            daily: d3.timeDay.every(1),
-            days: d3.timeDay.every(1),
-            hours: d3.timeHour.every(1),
-        }[intvl];
-    }
-    function getTicksMinor(intvl) {
-        return {
-            century: d3.timeYear.every(10),
-            jubilee: d3.timeYear.every(10),
-            decade: d3.timeYear.every(1),
-            lustrum: d3.timeYear.every(1),
-            years: d3.timeMonth.every(1),
-            fiscal: d3.timeMonth.every(1),
-            quarters: d3.timeMonth.every(3),
-            months: d3.timeDay.every(1),
-            weeks: d3.timeDay.every(1),
-            daily: d3.timeHour.every(1),
-            days: d3.timeHour.every(1),
-            hours: d3.timeMinute.every(1),
-        }[intvl];
-    }
-
-    function tickFormat(intvl) {
-        const formatFullYear = d3.timeFormat('%Y');
-        const formatYear = d3.timeFormat('%y');
-        const formatMonth = d3.timeFormat('%b');
-        const formatWeek = d3.timeFormat('%W');
-        const formatDay = d3.timeFormat('%d');
-        const formatHour = d3.timeFormat('%H:%M');
-        return {
-            century: d3.timeFormat('%Y'),
-            jubilee(d, i) {
-                const format = checkCentury(d, i);
-                return format;
-            },
-            decade(d, i) {
-                const format = checkCentury(d, i);
-                return format;
-            },
-            lustrum(d, i) {
-                const format = checkCentury(d, i);
-                return format;
-            },
-            years(d, i) {
-                const format = checkCentury(d, i);
-                return format;
-            },
-            fiscal(d, i) {
-                const format = getFiscal(d, i);
-                return format;
-            },
-            quarters(d, i) {
-                const format = getQuarters(d, i);
-                return format;
-            },
-            months(d, i) {
-                const format = checkMonth(d, i);
-                return format;
-            },
-            weeks(d, i) {
-                const format = getWeek(d, i);
-                return format;
-            },
-            days(d, i) {
-                const format = getDays(d, i);
-                return format;
-            },
-            daily(d, i) {
-                const format = getDaily(d, i);
-                return format;
-            },
-            hours(d, i) {
-                const format = getHours(d, i);
-                return format;
-            },
-        }[intvl];
-
-        function getHours(d, i) {
-            if (d.getHours() === 1 || i === 0) {
-                return `${formatHour(d)} ${formatDay(d)}`;
-            }
-            return formatHour(d);
-        }
-
-        function getDays(d, i) {
-            if (d.getDate() === 1 || i === 0) {
-                return `${formatDay(d)} ${formatMonth(d)}`;
-            }
-            return formatDay(d);
-        }
-
-        function getDaily(d, i) {
-            const last = scale.domain().length - 1;
-            if (i === 0) {
-                return `${formatDay(d)} ${formatMonth(d)}`;
-            }
-            if (d.getDate() === 1) {
-                return `${formatMonth(d)}`;
-            }
-            if (d.getDay() === 5) {
-                return `${formatDay(d)}`;
-            }
-            if (i === last) {
-                return formatDay(d);
-            }
-            return '';
-        }
-
-        function getWeek(d) {
-            if (d.getDate() < 9) {
-                return `${formatWeek(d)} ${formatMonth(d)}`;
-            }
-            return formatWeek(d);
-        }
-
-        function getQuarters(d, i) {
-            if (d.getMonth() < 3 && i < 4) {
-                return `Q1 ${formatFullYear(d)}`;
-            }
-            if (d.getMonth() < 3) {
-                return 'Q1';
-            }
-            if (d.getMonth() >= 3 && d.getMonth() < 6) {
-                return 'Q2';
-            }
-            if (d.getMonth() >= 6 && d.getMonth() < 9) {
-                return 'Q3';
-            }
-            if (d.getMonth() >= 9 && d.getMonth() < 12) {
-                return 'Q4';
-            }
-            throw new Error('Invalid quarter');
-        }
-
-        function checkMonth(d, i) {
-            if (d.getMonth() === 0 || i === 0) {
-                const newYear = d3.timeFormat('%b %Y');
-                return newYear(d);
-            }
-            return formatMonth(d);
-        }
-
-        function checkCentury(d, i) {
-            if (fullYear || +formatFullYear(d) % 100 === 0 || i === 0) {
-                return formatFullYear(d);
-            }
-            return formatYear(d);
-        }
-        function getFiscal(d, i) {
-            if (fullYear || +formatFullYear(d) % 100 === 0 || i === 0) {
-                return `${formatFullYear(d)}/${Number(formatYear(d)) + 1}`;
-            }
-            return `${formatYear(d)}/${Number(formatYear(d)) + 1}`;
-        }
-    }
     axis.align = (d) => {
         align = d;
         return axis;
