+++ src/yLinear.js
@@ -4,7 +4,12 @@
  */
 
 import * as d3 from 'd3';
-import { getDecimalFormat, getAxis } from './utils';
+import {
+    getAxis,
+    getDecimalFormat,
+    getDefaultYAxisLabel,
+    setLabelIds,
+} from './utils';
 
 export default function () {
     let banding;
@@ -105,22 +110,11 @@ export default function () {
         }
 
         if (frameName) {
-            yLabel
-                .selectAll('.axis.yAxis text')
-                .attr('id', `${frameName}yLabel`);
-            yLabel
-                .selectAll('.axis.yAxis line')
-                .attr('id', `${frameName}yTick`);
+            setLabelIds({ selection: yLabel, axis: 'y', frameName });
         }
 
         if (label) {
-            const defaultLabel = {
-                tag: label.tag,
-                hori: label.hori || 'left',
-                vert: label.vert || 'middle',
-                anchor: label.anchor || 'middle',
-                rotate: label.rotate || -90,
-            };
+            const defaultLabel = getDefaultYAxisLabel(label);
 
             const axisLabel = parent.append('g').attr('class', 'axis xAxis');
